requirements.txt
Flask
torch 
torchvision
gunicorn
numpy
pillow